requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
execnet==2.0.2
factory-boy==3.3.0
Faker==19.3.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
